Fixed LIR bug with return statements
Before, when `return` was used, the code generator would generate
assembly code which was identical to ending a function. This just
copied the return value over the arguments and returned. But, this
doesn't work for `return` statements, because they can return
arbitrarily far up the stack in the function.

So, now, the code generator sets the stack pointer to point to the
start of the stack frame after copying the return value, so that
when the function returns, it will leave the stack as intended.